Slightly improve login box phrasing
Honestly, the login box is horrible UX for newcomers, who get really
confused by the password prompt and think it's a register screen rather
than a login screen.

Phrasing won't help, but I honestly can't think of a good fix here...

The best way is probably to split the buttons into "New account" and
"Log in". This'll fix the confusion but make it take more work to log
in. _Probably_ worth it.