Add Gen 3 Pressure warning
It is possible, in most cases, to correctly track Pressure PP in Gen 3.
It is, however, a fuckton of work, and I'm way too busy to do that work
right now.

So instead, have a warning about how PP tracking might be wrong in the
event of Pressure.

TODO: Send silent still Pressure activations using |split|