Intercept large teams
PS the server will ignore all messages above ~150 KB. To protect against
DoS attacks, this is done silently, with no response at all.

Note that this is is not at all a problem for normal teams with 6
Pokemon, which come in at around 1KB.

Users sometimes make a mistake and send corrupt teams above this limit,
such as an import error causing a team with 1000 Pokemon or whatever.
This used to happen silently, confusing users and causing them to retry,
but this change gives it an error message.