Move EV suggester to battle-tooltips.ts
I mostly wanted this to be in TypeScript before I did any further work
on it, and also with the teambuilder at 3k lines, it needs to be broken
apart anyway.

I feel mildly bad about introducing ~600 lines of code that replays
will never need, but the EV suggester does fall under the general
category of "set insights" and is relevant to tooltips.