Only display one icon per species for Illusion
Previously, if Illusion caused two distinguishable copies of a Pokemon
to appear (either because they have different levels, or because they
switched into each other), the client would point this out by having
two Pokemon in the Pokemon list.

Due to this being by far the most commonly-reported "bug" in PS, this
has been revised to showing one copy in the team list, with the tooltip
showing details for every copy.

It will now be harder to tell that an oppsing Pokemon is an Illusion,
but oh well. I want to see more successful Zoroark usage, too, so this
isn't all bad.

This mostly affects Random Battle, which has the "different levels"
problem.