Allow data-send on <form> and <button>
This is slated to be the newest and easiest way to make buttons do
things:

`<button data-send="/join lobby">`

`<button name="send" value="/join lobby">`

Neither one is implemented yet, of course; that will come in a future
commit. This just allows them through the HTML sanitizer, which is
apparently the hardest part.