Build: Move ads/analytics into config
This keeps ads/analytics out of the repository, since they
shouldn't be used in forks and dev machines anyway.

Hiding ads in `website/news/include.php` was a thing I did in
the early days because one of the devs was ashamed of ads, and
`website/` wasn't in a public repository at the time. At this
point, though, I'm done being ashamed of it. You can obviously
see ads on the website and everyone knows they fund the server.