Battle: Support autoresize
PS's current desktop client supports mobile. But we currently use battle
scenes in four places:

- PS itself
- Replays
- Downloaded replays
- PS Preact client

And the other three don't support mobile. That changes today! Well, it
doesn't, but it changes sometime in the future, with this being the
next step.

So that the code doesn't need to be duplicated between all four clients,
it's been moved from `client-battle.js` to `battle.ts`.