Restore downloaded replay CSS
PR #2162 removed the CSS on downloaded replays without any
explanation. I think it was intended to save space and/or make
things bigger? But it came with a number of bugs.

- downloaded replay logs are much wider than intended
- downloaded replays now have a "battle options" button that does
  nothing
- downloaded replays have some unstyled text in e.g. popups
- MOST IMPORTANTLY: downloaded replays are no longer styled when
  you don't have an internet connection

Anyway, reverting that change fixes all these issues, and doesn't
seem to cause any new ones? This is a temporary measure while
waiting for Karthik to explain why the CSS removal was done in
the first place.

Refs #2165