Make sure replay downloading works correctly
Replays must be at their end before downloading replays works as
expected.

Before, fast-forwarding a battle would synchronously place replays
at their end, but it's now an async process as of
3ff296509fcbe9a39be3c6f

Telling the user to click the Download button again is very hacky,
but this is a very rare situation not really worth the effort to
make it more elegant.