Preact: Open console when pressing / in Main Menu
There's no easy way to enter a command in PS if you're not logged in.
This adds one.

This also standardizes the room ID for the console as `pm-`.

I'm still mulling whether `pm-[userid1]-[userid2]` should be the roomid,
or `pm-[pmtarget]`. Arguably it's nice for PMs to have the same roomid
for both sides of the conversation, but also arguably it's nice for
PMs not to change roomid when you change your own name (it'd still
change if your PM target changed name).