Preact: Pass clicked button to commands
Because commands are frequently called from buttons as well as
textboxes, passing the clicked button is useful.

The textbox is also passed if the command is typed rather than from
a button. Mostly this just lets us spawn popups closer to where users
are already looking, which I've been wanting to do for `/nick` for
a while.