Improve support for unregistered servers
Client-side of https://github.com/smogon/pokemon-showdown/commit/e7b2be606aeb1b6387140c9de6ed440a5ea51d20

Unregistered server URLs are now pretty different (see the source
of the link). Part of this is because of some weird change to the
web server config, part of this is because HTTPS wildcard certs don't
work infinitely deep, part of this is because Chrome now insists
on forcing HTTPS in a variety of scenarios...

Anyway, I'd love to keep supporting the old syntax but it's not
workable, and I'm really happy to have a new one that is.